Now use SD card slot 1 or 2 and download DNG or JPG.
Main changes allow user to select to download files from card slot 1 or 2, and to choose either DNG or JPG.  By default JPG from card 2 are no downloaded. 
Code was rearranged, slightly better commented, and refactored to be more separate form the original FluCard based code.
The basic workflow is the same and is rather clunky. i.e. if you have more than about 500 photos on the card, the overhead can cause lost communication and require you re-run the program.    I hope to address this issue in the next version.